using CSVLoader() instead to be more efficient with runtime
