+++ backend/requirements.txt
@@ -9,3 +9,4 @@ langchain==0.3.12
 langchain-openai==0.2.12
 langchain-chroma==0.1.4
 requests==2.32.3
+langchain-community==0.3.12
